refactor($function): refactor functions and split functions to other files to make it clean
