+++ src/Host.js
@@ -1,0 +1,3 @@
+const HOST = "https://api.github.com";
+
+export {HOST};
