+++ src/render.js
@@ -1,0 +1,12 @@
+const renderHtml = ownerAndFollowers => {
+    const followerAvatars = ownerAndFollowers.followers.map(
+        follower => follower.avatar_url
+    );
+
+    document.write(`<li>${ownerAndFollowers.owner}</li>`);
+    followerAvatars.map(avatar =>
+        document.write(`<img src='${avatar}' height='42' width='42'>`)
+    );
+};
+
+export default renderHtml;
